==> explore/Mapping_process/2_COVID_h5_to_SeuratObject_merge.Rmd <==
---
title: "COVID_h5_to_SeuratObject_merge"
output: html_document
output: github_document
---


```{r setup, include=FALSE}
library(Matrix)
library(rhdf5)
library(tidyverse)
library(glue)
library(here)
library(hdf5r)
```

```{r}
# C166
data <- Read10X_h5(here('data/GSE149878_covid_lung/GSM4516279_C166_filtered_feature_bc_matrix.h5'), use.names = TRUE, unique.features = TRUE)
obj1 <- CreateSeuratObject(data)
obj1
```

```{r}
# C168
data <- Read10X_h5(here('data/GSE149878_covid_lung/GSM4516280_C168_filtered_feature_bc_matrix.h5'), use.names = TRUE, unique.features = TRUE)
obj2 <- CreateSeuratObject(data)
obj2
```

```{r}
# C170
data <- Read10X_h5(here('data/GSE149878_covid_lung/GSM4516281_C170_filtered_feature_bc_matrix.h5'), use.names = TRUE, unique.features = TRUE)
obj3 <- CreateSeuratObject(data)
obj3
```

```{r}
# C172
data <- Read10X_h5(here('data/GSE149878_covid_lung/GSM4516282_C172_filtered_feature_bc_matrix.h5'), use.names = TRUE, unique.features = TRUE)
obj4 <- CreateSeuratObject(data)
obj4
```

```{r}
# merge SeuratObject
covid <- merge(obj1, y = c(obj2, obj3, obj4), add.cell.ids = c('obj1', 'obj2', 'obj3', 'obj4'), project = 'scRNA_IPF_data')
covid
```

```{r}
head(covid)
```

```{r}
unique(sapply(X = strsplit(colnames(covid), split = "_"), FUN = "[", 1))
```

```{r}
saveRDS(covid, file = "COVID_h5_to_SeuratObject_merge.RDS")
```
