==> explore/Mapping_process/9_COVID_github_replication.Rmd <==
---
title: "COVID_github_replication"
output: html_document
output: github_document
---

```{r setup, include=FALSE}
library(Seurat)
library(here)
library(tidyverse)
library(cowplot)
library(knitr)
library(BiocManager)
library(sctransform)
library(glmGamPoi)
library(Matrix)

library(Seurat)
library(dplyr)
library(ggplot2)
library(ade4)
library(Matrix)
set.seed(8211)
```

```{r}
covid <- readRDS(here('explore/RDS/COVID_Preprocessing.RDS'))
head(covid)
```

```{r}
# Chose a PC for analysis
covid <- FindVariableFeatures(covid, verbose = T, nfeatures = 3000)
covid <- ScaleData(covid, features = row.names(covid@assays$SCT@data))
covid <- RunPCA(covid)

potato <-  matrix(ncol = 17, nrow = ncol(covid))
potato2 <-  matrix(ncol = 17, nrow = ncol(covid))

j <- 0
for(i in 9:25) {
  j <- j + 1
  print(i)
  onion <- RunUMAP(covid, dims = 1:i, verbose = F)
  potato[, j] <- onion@reductions$umap@cell.embeddings[, 1]
  potato2[, j] <- onion@reductions$umap@cell.embeddings[, 2]
}

colnames(potato) <- 9:25
colnames(potato2) <- 9:25
mrand_obs <- NULL
potato_new <- potato[sample(1:ncol(covid), 3000), ]
potato2_new <- potato2[sample(1:ncol(covid), 3000), ]

for(i in  1:c(dim(potato_new)[2] - 1)) {
  onion <- dist(potato_new[, i])
  onion2 <- dist(potato_new[, i + 1])
  onion3 <- mantel.randtest(onion, onion2)$obs
  onion <- dist(potato2_new[, i])
  onion2 <- dist(potato2_new[, i + 1])
  onion4 <- mantel.randtest(onion, onion2)$obs
  mrand_obs <- rbind(mrand_obs, cbind(onion3, onion4))
}

pdf("covid_PCA.pdf")
plot(c(9,10,11,12,13,14,15,16,17,18,19,20,21,22,23,24),
     mrand_obs[, 1], ylim = c(0.8, 1))
lines(spline(x = c(9,10,11,12,13,14,15,16,17,18,19,20,21,22,23,24),
             y = mrand_obs[, 1]))
points(c(9,10,11,12,13,14,15,16,17,18,19,20,21,22,23,24),
       mrand_obs[, 2], ylim = c(0.8, 1), col = "red")
lines(spline(x = c(9,10,11,12,13,14,15,16,17,18,19,20,21,22,23,24),
             y = mrand_obs[, 2]), col = "red")

ElbowPlot(covid)
dev.off()
rm(onion,onion2,onion3,onion4,mrand_obs,potato,potato2, potato_new,potato2_new)


covid <- RunUMAP(object = covid, dims = 1:20, verbose = F)
covid <- FindNeighbors(object = covid, dims = 1:20, verbose = F)
covid <- FindClusters(object = covid, resolution = 0.01, verbose = F)
```
 
 
```{r}
saveRDS(covid, file = "COVID_github_replication.RDS")
```
 
 
```{r}
pdf("covid.pdf")
DimPlot(covid)
FeaturePlot(covid, c("PTPRC", "EPCAM", "PECAM1"))
DotPlot(covid, features = c("PTPRC", "EPCAM", "PECAM1"))
dev.off()

Idents(covid, cells = WhichCells(covid, idents = c(0, 3))) <- "Immune"
Idents(covid, cells = WhichCells(covid, idents = c(1, 2))) <- "Epithelial"
Idents(covid, cells = WhichCells(covid, idents = c(4, 6))) <- "Endothelial"
Idents(covid, cells = WhichCells(covid, idents = c(5))) <- "Mesenchymal"
covid@meta.data$population <- Idents(covid)


unique(covid@meta.data$population)
table(is.na(covid@meta.data$population))


# Subset into 4 major populations
# ======================================
immune <- subset(covid, idents = "Immune")
epi <- subset(covid, idents = "Epithelial")
endo <- subset(covid, idents = "Endothelial")
meso <- subset(covid, idents = "Mesenchymal")

# Confirm subsetted data
table(covid@meta.data$population)
table(immune@meta.data$population)
table(epi@meta.data$population)
table(endo@meta.data$population)
table(meso@meta.data$population)
```

```{r}
ncol(meso)
```


```{r}
# ======================================
# Make plots to determine which PC 
# to use for each subset
# Mesenchymal subset
# ======================================
meso <- FindVariableFeatures(meso, verbose = T, nfeatures = 3000)
meso <- ScaleData(meso,features = row.names(meso@assays$SCT@data))
meso <- RunPCA(meso)

potato <- matrix(ncol = 20, nrow = ncol(meso))
potato2 <- matrix(ncol = 20, nrow = ncol(meso))

j <- 0
for(i in 6:25) {
  j <- j + 1
  print(i)
  onion <- RunUMAP(meso, dims = 1:i, verbose = F)
  potato[, j] <- onion@reductions$umap@cell.embeddings[, 1]
  potato2[, j] <- onion@reductions$umap@cell.embeddings[, 2]
}



colnames(potato) <- 6:25
colnames(potato2) <- 6:25
mrand_obs <- NULL
potato_new <- potato[sample(1:ncol(meso), 10000]#, replace = TRUE, prob = NULL)]
potato2_new <- potato2[sample(1:ncol(meso), 10000]#, replace = TRUE, prob = NULL)]

```




```{r}
for(i in  1:c(dim(potato_new)[2] - 1)) {
  onion <- dist(potato_new[, i])
  onion2 <- dist(potato_new[, i + 1])
  onion3 <- mantel.randtest(onion, onion2)$obs
  onion <- dist(potato2_new[, i])
  onion2 <- dist(potato2_new[, i + 1])
  onion4 <- mantel.randtest(onion, onion2)$obs
  mrand_obs <- rbind(mrand_obs, cbind(onion3, onion4))}

pdf("meso_PCA.pdf")
plot(c(6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21,22,23,24),
     mrand_obs[, 1], ylim = c(0, 1))
lines(spline(x = c(6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21,22,23,24),
             y = mrand_obs[, 1]))
points(c(6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21,22,23,24),
       mrand_obs[, 2], ylim = c(0, 1), col = "red")
lines(spline(x = c(6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21,22,23,24),
             y = mrand_obs[, 2]), col = "red")

dev.off()
rm(onion,onion2,onion3,onion4,mrand_obs,potato,potato2, potato_new,potato2_new)
```

```{r}
# ======================================
# Endothelial subset
# ======================================
endo <- FindVariableFeatures(endo, verbose = T, nfeatures = 3000)
endo <- ScaleData(endo,features = row.names(endo@assays$SCT@data))
endo <- RunPCA(endo)
potato <- matrix(ncol = 12, nrow = ncol(endo))
potato2 <- matrix(ncol = 12, nrow = ncol(endo))

j <- 0
for(i in 9:20) {
  j <- j + 1
  print(i)
  onion <- RunUMAP(endo, dims = 1:i, verbose = F)
  potato[, j] <- onion@reductions$umap@cell.embeddings[, 1]
  potato2[, j] <- onion@reductions$umap@cell.embeddings[, 2]
}

colnames(potato) <- 9:20
colnames(potato2) <- 9:20
mrand_obs <- NULL
potato_new <- potato[sample(1:ncol(endo), 3000),replace = True]
potato2_new <- potato2[sample(1:ncol(endo), 3000),replace = True]

for(i in  1:c(dim(potato_new)[2] - 1)) {
  onion <- dist(potato_new[, i])
  onion2 <- dist(potato_new[, i + 1])
  onion3 <- mantel.randtest(onion, onion2)$obs
  onion <- dist(potato2_new[, i])
  onion2 <- dist(potato2_new[, i + 1])
  onion4 <- mantel.randtest(onion, onion2)$obs
  mrand_obs <- rbind(mrand_obs, cbind(onion3, onion4))
}

pdf("endo_PCA.pdf")
plot(c(9,10,11,12,13,14,15,16,17,18,19),
     mrand_obs[, 1], ylim = c(.75, 1))
lines(spline(x = c(9,10,11,12,13,14,15,16,17,18,19),
             y = mrand_obs[, 1]))
points(c(9,10,11,12,13,14,15,16,17,18,19),
       mrand_obs[, 2], ylim = c(.75, 1), col = "red")
lines(spline(x = c(9,10,11,12,13,14,15,16,17,18,19),
             y = mrand_obs[, 2]), col = "red")

ElbowPlot(endo)
dev.off()
rm(onion,onion2,onion3,onion4,mrand_obs,potato,potato2, potato_new,potato2_new)

# ======================================
# Epithelial subset
# ======================================
epi <- FindVariableFeatures(epi, verbose = T, nfeatures = 3000)
epi <- ScaleData(epi,features = row.names(epi@assays$SCT@data))
epi <- RunPCA(epi)
potato <- matrix(ncol = 15, nrow = ncol(epi))
potato2 <- matrix(ncol = 15, nrow = ncol(epi))

j <- 0
for(i in 7:21) {
  j <- j + 1
  print(i)
  onion <- RunUMAP(epi, dims = 1:i, verbose = F)
  potato[, j] <- onion@reductions$umap@cell.embeddings[, 1]
  potato2[, j] <- onion@reductions$umap@cell.embeddings[, 2]
}

colnames(potato) <- 7:21
colnames(potato2) <- 7:21
mrand_obs <- NULL
potato_new <- potato[sample(1:ncol(epi), 3000),]
potato2_new <- potato2[sample(1:ncol(epi), 3000),]

for(i in  1:c(dim(potato_new)[2] - 1)) {
  onion <- dist(potato_new[, i])
  onion2 <- dist(potato_new[, i + 1])
  onion3 <- mantel.randtest(onion, onion2)$obs
  onion <- dist(potato2_new[, i])
  onion2 <- dist(potato2_new[, i + 1])
  onion4 <- mantel.randtest(onion, onion2)$obs
  mrand_obs <- rbind(mrand_obs, cbind(onion3, onion4))
}

pdf("epi_PCA.pdf")
plot(c(7,8,9,10,11,12,13,14,15,16,17,18,19,20),
     mrand_obs[, 1], ylim = c(.75, 1))
lines(spline(x = c(7,8,9,10,11,12,13,14,15,16,17,18,19,20),
             y = mrand_obs[, 1]))
points(c(7,8,9,10,11,12,13,14,15,16,17,18,19,20),
       mrand_obs[, 2], ylim = c(.75, 1), col = "red")
lines(spline(x = c(7,8,9,10,11,12,13,14,15,16,17,18,19,20),
             y = mrand_obs[, 2]), col = "red")

ElbowPlot(epi, ndims = 30)
dev.off()
rm(onion,onion2,onion3,onion4,mrand_obs,potato,potato2, potato_new,potato2_new)

# ======================================
# Immune subset
# ======================================
immune <- FindVariableFeatures(immune, verbose = T, nfeatures = 3000)
immune <- ScaleData(immune, features = row.names(immune@assays$SCT@data))
immune <- RunPCA(immune)
potato <- matrix(ncol = 15, nrow = ncol(immune))
potato2 <- matrix(ncol = 15, nrow = ncol(immune))

j <- 0
for(i in 13:27) {
  j <- j + 1
  print(i)
  onion <- RunUMAP(immune, dims = 1:i, verbose = F)
  potato[, j] <- onion@reductions$umap@cell.embeddings[, 1]
  potato2[, ] <- onion@reductions$umap@cell.embeddings[, 2]
}

colnames(potato) <- 13:27
colnames(potato2) <- 13:27
mrand_obs <- NULL
potato_new <- potato[sample(1:ncol(immune), 3000),]
potato2_new <- potato2[sample(1:ncol(immune), 3000),]

for(i in  1:c(dim(potato_new)[2] - 1)){
  onion <- dist(potato_new[, i])
  onion2 <- dist(potato_new[, i + 1])
  onion3 <- mantel.randtest(onion, onion2)$obs
  onion <- dist(potato2_new[, i])
  onion2 <- dist(potato2_new[, i + 1])
  onion4 <- mantel.randtest(onion, onion2)$obs
  mrand_obs <- rbind(mrand_obs, cbind(onion3, onion4))
}

pdf("immune_PCA.pdf")
plot(c(13,14,15,16,17,18,19,20,21,22,23,24,25,26),
     mrand_obs[, 1], ylim = c(0, 1))
lines(spline(x = c(13,14,15,16,17,18,19,20,21,22,23,24,25,26),
             y = mrand_obs[, 1]))
points(c(13,14,15,16,17,18,19,20,21,22,23,24,25,26),
       mrand_obs[, 2], ylim = c(0, 1), col = "red")
lines(spline(x = c(13,14,15,16,17,18,19,20,21,22,23,24,25,26),
             y = mrand_obs[, 2]), col = "red")

ElbowPlot(immune, ndims = 30)
dev.off()
```

```{r}
epi <- RunUMAP(object = epi, dims = 1:11, verbose = F)
endo <- RunUMAP(object = endo, dims = 1:14, verbose = F)
immune <- RunUMAP(object = immune, dims = 1:21, verbose = F)
meso <- RunUMAP(object = meso, dims = 1:14, verbose = F)
```

```{r}
runFeaturePlots = TRUE
epi <- FindNeighbors(epi, dims = 1:20)
epi <- FindClusters(epi, resolution = 2)

if (runFeaturePlots) {FeaturePlot(epi, features = c("SFTPC", "ABCA3", "AGER", "HOPX"))}
Idents(epi, cells = WhichCells(epi, idents = c(19))) <- "AT1"
Idents(epi, cells = WhichCells(epi, idents = c(3, 9, 10, 13, 18, 22, 25, 28, 29))) <- "AT2"
Idents(epi, cells = WhichCells(epi, idents = c(17, 31))) <- "Transitional AT2"

if (runFeaturePlots) {FeaturePlot(epi, features = c("KRT5", "KRT17", "TP63", "COL1A1"))}
Idents(epi, cells = WhichCells(epi, idents = c(6))) <- "Basal"
Idents(epi, cells = WhichCells(epi, idents = c(23))) <- "KRT5-/KRT17+"

if (runFeaturePlots) {FeaturePlot(epi, features = c("SCGB1A1", "SCGB3A2", "MUC5B", "MUC5AC"))}
Idents(epi, cells = WhichCells(epi, idents = c(8, 20))) <- "MUC5B+"
Idents(epi, cells = WhichCells(epi, idents = c(0, 24))) <- "SCGB3A2+"
if (runFeaturePlots) {FeaturePlot(epi, features = c("MGP", "SCGB1A1", "SCGB3A2"))}
Idents(epi, cells = WhichCells(epi, idents = c(12))) <- "SCGB3A2+ SCGB1A1+"

if (runFeaturePlots) {FeaturePlot(epi, features = c("FOXI1", "CHGA"))}
Idents(epi, cells = WhichCells(epi, idents = c(38))) <- "PNECs/Ionocytes"

if (runFeaturePlots) {FeaturePlot(epi, features = c("FOXJ1", "TMEM190", "CAPS", "HYDIN"))}
Idents(epi, cells = WhichCells(epi, idents = c(1, 2, 4, 5, 7, 11, 16, 21, 27, 35))) <- "Ciliated"
Idents(epi, cells = WhichCells(epi, idents = c(14, 34))) <- "Differentiating Ciliated"

if (runFeaturePlots) {FeaturePlot(epi, features = c("EPCAM", "PTPRC", "PECAM1", "LUM"))}
if (runFeaturePlots) {FeaturePlot(epi, features = c("SFTPC", "SCGB1A1", "SCGB3A2", "FOXJ1", "KRT5", "AGER"))}
Idents(epi, cells = WhichCells(epi, idents = c(15, 26, 32, 33, 36, 37))) <- "Doublets"

if (runFeaturePlots) {FeaturePlot(epi, features = c("MKI67"))}
Idents(epi, cells = WhichCells(epi, idents = c(30))) <- "Proliferating Epithelial Cells"

epi$firstCT <- Idents(epi)

pnecionocyte_subset <- subset(epi, idents = c("PNECs/Ionocytes"))
if (runFeaturePlots) {FeaturePlot(pnecionocyte_subset, features = c("FOXI1", "CHGA"))}
Idents(epi, cells = WhichCells(pnecionocyte_subset, expression = CHGA > 0)) <- "PNECs"
Idents(epi, cells = WhichCells(pnecionocyte_subset, expression = FOXI1 > 0)) <- "Ionocytes"
Idents(epi, cells = WhichCells(pnecionocyte_subset, idents = "PNECs/Ionocytes")) <- "Basal"

muc5b_subset <- subset(epi, idents = c("MUC5B+")) %>% FindNeighbors() %>% FindClusters(resolution = 2)
if (runFeaturePlots) {FeaturePlot(muc5b_subset, features = c("MUC5B", "MUC5AC"))}
Idents(epi, cells = WhichCells(muc5b_subset, idents = c(15))) <- "MUC5AC+ High"

epi$celltype <- Idents(epi)
DimPlot(epi)

keep <- c("AT2","Differentiating Ciliated","PNECs","MUC5AC+ High","Transitional AT2", "SCGB3A2+ SCGB1A1+", "AT1", "Ionocytes", "MUC5B+","SCGB3A2+","Basal","Ciliated","KRT5-/KRT17+", "Proliferating Epithelial Cells")
epi <- subset(epi, cells = row.names(epi@meta.data[epi@meta.data$celltype %in% keep, ]))
DimPlot(epi, group.by = "celltype")

potato <- matrix(ncol = 15, nrow = ncol(epi))
potato2 <- matrix(ncol = 15, nrow = ncol(epi))

j <- 0
for(i in 7:21) {
  j <- j + 1
  print(i)
  onion <- RunUMAP(epi, dims = 1:i, verbose = F)
  potato[, j] <- onion@reductions$umap@cell.embeddings[, 1]
  potato2[, j] <- onion@reductions$umap@cell.embeddings[, 2]
}

colnames(potato) <- 7:21
colnames(potato2) <- 7:21
mrand_obs <- NULL
potato_new <- potato[sample(1:ncol(epi), 3000),]
potato2_new <- potato2[sample(1:ncol(epi), 3000),]

for(i in  1:c(dim(potato_new)[2] - 1)) {
  onion <- dist(potato_new[, i])
  onion2 <- dist(potato_new[, i + 1])
  onion3 <- mantel.randtest(onion, onion2)$obs
  onion <- dist(potato2_new[, i])
  onion2 <- dist(potato2_new[, i + 1])
  onion4 <- mantel.randtest(onion, onion2)$obs
  mrand_obs <- rbind(mrand_obs, cbind(onion3, onion4))
}

pdf("epi_PCA.pdf")
plot(c(7,8,9,10,11,12,13,14,15,16,17,18,19,20),
     mrand_obs[, 1], ylim = c(.75, 1))
lines(spline(x = c(7,8,9,10,11,12,13,14,15,16,17,18,19,20),
             y = mrand_obs[, 1]))
points(c(7,8,9,10,11,12,13,14,15,16,17,18,19,20),
       mrand_obs[, 2], ylim = c(.75, 1), col = "red")
lines(spline(x = c(7,8,9,10,11,12,13,14,15,16,17,18,19,20),
             y = mrand_obs[, 2]), col = "red")

ElbowPlot(epi, ndims = 30)
dev.off()

epi <- RunUMAP(epi, dims = 1:11)
DimPlot(epi, group.by = "celltype")

# ======================================
# Immune
# ======================================
immune <- FindNeighbors(immune, dims = 1:20)
immune <- FindClusters(immune, resolution = 1)

if (runFeaturePlots) {FeaturePlot(immune, features = c("LYZ", "MARCO", "FCGR1A", "C1QB"))}
Idents(immune, cells = WhichCells(immune, idents = c(0, 1, 2, 3, 4, 7, 8, 9, 11, 12, 14, 19, 25, 30))) <- "Macrophages"

if (runFeaturePlots) {FeaturePlot(immune, features = c("CD3E", "FOXP3", "IL7R", "CD8A"))}
Idents(immune, cells = WhichCells(immune, idents = c(5, 15, 23, 27, 29))) <- "T Cells"

if (runFeaturePlots) {FeaturePlot(immune, features = c("CD14", "LYZ", "S100A12", "ITGAL"))}
Idents(immune, cells = WhichCells(immune, idents = c(6, 18, 20, 21))) <- "Monocytes"

if (runFeaturePlots) {FeaturePlot(immune, features = c("MS4A1", "CD19", "JCHAIN", "IGHG1"))}
Idents(immune, cells = WhichCells(immune, idents = c(22))) <- "B Cells"
Idents(immune, cells = WhichCells(immune, idents = c(26))) <- "Plasma Cells"

if (runFeaturePlots) {FeaturePlot(immune, features = c("CPA3", "KIT"))}
Idents(immune, cells = WhichCells(immune, idents = c(24))) <- "Mast Cells"

if (runFeaturePlots) {FeaturePlot(immune, features = c("FCER1A", "CD1C", "CLEC9A", "FSCN1"))}
Idents(immune, cells = WhichCells(immune, idents = c(13))) <- "cDCs"
if (runFeaturePlots) {FeaturePlot(immune, features = c("LILRA4", "CLEC4C", "JCHAIN", "TCF4"))}
Idents(immune, cells = WhichCells(immune, idents = c(31))) <- "pDCs"

if (runFeaturePlots) {FeaturePlot(immune, features = c("CD3E", "KLRB1", "NKG7", "NCR1"))}
Idents(immune, cells = WhichCells(immune, idents = c(10))) <- "NK Cells" 

if (runFeaturePlots) {FeaturePlot(immune, features = c("MKI67", "CD3E", "LYZ"))}
if (runFeaturePlots) {FeaturePlot(immune, features = c("PTPRC", "EPCAM", "VWF", "ACTA2"))}
Idents(immune, cells = WhichCells(immune, idents = c(28))) <- "Proliferating T Cells"
Idents(immune, cells = WhichCells(immune, idents = c(16))) <- "Proliferating Macrophages"
Idents(immune, cells = WhichCells(immune, idents = c(17))) <- "Doublets"

immune$celltype <- Idents(immune)
DimPlot(immune)

DimPlot(immune, group.by = "celltype")
immune_keep <- c("cDCs", "Proliferating Macrophages", "Proliferating T Cells", "NK Cells", "Plasma Cells", "moDCs", "Mast Cells", "B Cells", "Monocytes", "pDCs", "T Cells", "Macrophages")
immune <- subset(immune, cells = row.names(immune@meta.data[immune@meta.data$celltype %in% immune_keep, ]))

potato <- matrix(ncol = 15, nrow = ncol(immune))
potato2 <- matrix(ncol = 15, nrow = ncol(immune))

j <- 0
for(i in 13:27) {
  j <- j + 1
  print(i)
  onion <- RunUMAP(immune, dims = 1:i, verbose = F)
  potato[, j] <- onion@reductions$umap@cell.embeddings[, 1]
  potato2[, ] <- onion@reductions$umap@cell.embeddings[, 2]
}

colnames(potato) <- 13:27
colnames(potato2) <- 13:27
mrand_obs <- NULL
potato_new <- potato[sample(1:ncol(immune), 3000),]
potato2_new <- potato2[sample(1:ncol(immune), 3000),]

for(i in  1:c(dim(potato_new)[2] - 1)){
  onion <- dist(potato_new[, i])
  onion2 <- dist(potato_new[, i + 1])
  onion3 <- mantel.randtest(onion, onion2)$obs
  onion <- dist(potato2_new[, i])
  onion2 <- dist(potato2_new[, i + 1])
  onion4 <- mantel.randtest(onion, onion2)$obs
  mrand_obs <- rbind(mrand_obs, cbind(onion3, onion4))
}

pdf("immune_PCA.pdf")
plot(c(13,14,15,16,17,18,19,20,21,22,23,24,25,26),
     mrand_obs[, 1], ylim = c(0, 1))
lines(spline(x = c(13,14,15,16,17,18,19,20,21,22,23,24,25,26),
             y = mrand_obs[, 1]))
points(c(13,14,15,16,17,18,19,20,21,22,23,24,25,26),
       mrand_obs[, 2], ylim = c(0, 1), col = "red")
lines(spline(x = c(13,14,15,16,17,18,19,20,21,22,23,24,25,26),
             y = mrand_obs[, 2]), col = "red")

ElbowPlot(immune, ndims = 30)
dev.off()
rm(onion,onion2,onion3,onion4,mrand_obs,potato,potato2, potato_new,potato2_new)

immune <- RunUMAP(immune, dims = 1:20)
DimPlot(immune, group.by = "celltype")

# ======================================
# Mesenchymal
# ======================================
meso <- FindNeighbors(meso, dims = 1:20)
meso <- FindClusters(meso, resolution = 0.5)

if (runFeaturePlots) {FeaturePlot(meso, features = c("LUM", "ACTA2", "WT1", "PDGFRA"))}
Idents(meso, cells = WhichCells(meso, idents = c(2, 3))) <- "Smooth Muscle Cells"
Idents(meso, cells = WhichCells(meso, idents = c(7))) <- "Mesothelial Cells"

if (runFeaturePlots) {FeaturePlot(meso, features = c("LUM", "HAS1", "ACTA2", "PLIN2"))}
Idents(meso, cells = WhichCells(meso, idents = c(1, 5, 6, 12))) <- "Myofibroblasts"
Idents(meso, cells = WhichCells(meso, idents = c(9))) <- "HAS1 High Fibrboblasts"
Idents(meso, cells = WhichCells(meso, idents = c(4))) <- "Fibroblasts"
Idents(meso, cells = WhichCells(meso, idents = c(0, 13))) <- "PLIN2+ Fibroblasts"

if (runFeaturePlots) {FeaturePlot(meso, features = c("PTPRC", "EPCAM", "PECAM1", "VWF"))}
Idents(meso, cells = WhichCells(meso, idents = c(8, 10, 11, 14))) <- "Doublets"

meso$celltype <- Idents(meso)
DimPlot(meso)

DimPlot(meso, group.by = "celltype")
onion <- as.character(meso@meta.data$celltype)
onion[onion == 'HAS1 High Fibrboblasts'] <- "HAS1 High Fibroblasts"
meso@meta.data$celltype <- onion
meso_keep <- c("PLIN2+ Fibroblasts", "Myofibroblasts", "HAS1 High Fibroblasts", "Mesothelial Cells", "Smooth Muscle Cells", "Fibroblasts")
meso <- subset(meso, cells = row.names(meso@meta.data[meso@meta.data$celltype %in% meso_keep, ]))

potato <- matrix(ncol = 20, nrow = ncol(meso))
potato2 <- matrix(ncol = 20, nrow = ncol(meso))

j <- 0
for(i in 6:25) {
  j <- j + 1
  print(i)
  onion <- RunUMAP(meso, dims = 1:i, verbose = F)
  potato[, j] <- onion@reductions$umap@cell.embeddings[, 1]
  potato2[, j] <- onion@reductions$umap@cell.embeddings[, 2]
}

colnames(potato) <- 6:25
colnames(potato2) <- 6:25
mrand_obs <- NULL
potato_new <- potato[sample(1:ncol(meso), 3000), ]
potato2_new <- potato2[sample(1:ncol(meso), 3000), ]

for(i in  1:c(dim(potato_new)[2] - 1)) {
  onion <- dist(potato_new[, i])
  onion2 <- dist(potato_new[, i + 1])
  onion3 <- mantel.randtest(onion, onion2)$obs
  onion <- dist(potato2_new[, i])
  onion2 <- dist(potato2_new[, i + 1])
  onion4 <- mantel.randtest(onion, onion2)$obs
  mrand_obs <- rbind(mrand_obs, cbind(onion3, onion4))
}

pdf("meso_PCA.pdf")
plot(c(6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21,22,23,24),
     mrand_obs[, 1], ylim = c(0, 1))
lines(spline(x = c(6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21,22,23,24),
             y = mrand_obs[, 1]))
points(c(6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21,22,23,24),
       mrand_obs[, 2], ylim = c(0, 1), col = "red")
lines(spline(x = c(6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21,22,23,24),
             y = mrand_obs[, 2]), col = "red")

dev.off()
rm(onion,onion2,onion3,onion4,mrand_obs,potato,potato2, potato_new,potato2_new)

meso <- RunUMAP(meso, dims = 1:16)
DimPlot(meso, group.by = "celltype")

# ======================================
# Endothelial
# ======================================
endo <- FindNeighbors(endo, dims = 1:10)
endo <- FindClusters(endo)

if (runFeaturePlots) {FeaturePlot(endo, features = c("VWF", "EDNRB", "CCL21", "PECAM1"))}
if (runFeaturePlots) {FeaturePlot(endo, features = c("PTPRC", "EPCAM", "LUM", "ACTA2"))}
Idents(endo, cells = WhichCells(endo, idents = c(0:6, 8, 9, 12:15, 18))) <- "Endothelial Cells"
Idents(endo, cells = WhichCells(endo, idents = c(7, 11))) <- "Lymphatic Endothelial Cells"
Idents(endo, cells = WhichCells(endo, idents = c(10, 16, 17))) <- "Doublets"
endo$celltype <- Idents(endo)

DimPlot(endo, group.by = "celltype")
endo_keep <- c("Lymphatic Endothelial Cells", "Endothelial Cells")
endo <- subset(endo, cells = row.names(endo@meta.data[endo@meta.data$celltype %in% endo_keep, ]))

potato <- matrix(ncol = 12, nrow = ncol(endo))
potato2 <- matrix(ncol = 12, nrow = ncol(endo))

j <- 0
for(i in 9:20) {
  j <- j + 1
  print(i)
  onion <- RunUMAP(endo, dims = 1:i, verbose = F)
  potato[, j] <- onion@reductions$umap@cell.embeddings[, 1]
  potato2[, j] <- onion@reductions$umap@cell.embeddings[, 2]
}

colnames(potato) <- 9:20
colnames(potato2) <- 9:20
mrand_obs <- NULL
potato_new <- potato[sample(1:ncol(endo), 3000),]
potato2_new <- potato2[sample(1:ncol(endo), 3000),]

for(i in  1:c(dim(potato_new)[2] - 1)) {
  onion <- dist(potato_new[, i])
  onion2 <- dist(potato_new[, i + 1])
  onion3 <- mantel.randtest(onion, onion2)$obs
  onion <- dist(potato2_new[, i])
  onion2 <- dist(potato2_new[, i + 1])
  onion4 <- mantel.randtest(onion, onion2)$obs
  mrand_obs <- rbind(mrand_obs, cbind(onion3, onion4))
}

pdf("endo_PCA.pdf")
plot(c(9,10,11,12,13,14,15,16,17,18,19),
     mrand_obs[, 1], ylim = c(.75, 1))
lines(spline(x = c(9,10,11,12,13,14,15,16,17,18,19),
             y = mrand_obs[, 1]))
points(c(9,10,11,12,13,14,15,16,17,18,19),
       mrand_obs[, 2], ylim = c(.75, 1), col = "red")
lines(spline(x = c(9,10,11,12,13,14,15,16,17,18,19),
             y = mrand_obs[, 2]), col = "red")

ElbowPlot(endo)
dev.off()
rm(onion,onion2,onion3,onion4,mrand_obs,potato,potato2, potato_new,potato2_new)

endo <- RunUMAP(endo, dims = 1:9)
DimPlot(endo, group.by = "celltype")

# ======================================
# Merge all objects back into covid
# ======================================
covid <- merge(x = epi, y=c(endo, meso, immune))
covid <- FindVariableFeatures(covid, verbose = T, nfeatures = 3000)
covid <- ScaleData(covid, features = row.names(covid@assays$SCT@data))
covid <- RunPCA(covid)

potato <-  matrix(ncol = 17, nrow = ncol(covid))
potato2 <-  matrix(ncol = 17, nrow = ncol(covid))

j <- 0
for(i in 9:25) {
  j <- j + 1
  print(i)
  onion <- RunUMAP(covid, dims = 1:i, verbose = F)
  potato[, j] <- onion@reductions$umap@cell.embeddings[, 1]
  potato2[, j] <- onion@reductions$umap@cell.embeddings[, 2]
}

colnames(potato) <- 9:25
colnames(potato2) <- 9:25
mrand_obs <- NULL
potato_new <- potato[sample(1:ncol(covid), 3000), ]
potato2_new <- potato2[sample(1:ncol(covid), 3000), ]

for(i in  1:c(dim(potato_new)[2] - 1)) {
  onion <- dist(potato_new[, i])
  onion2 <- dist(potato_new[, i + 1])
  onion3 <- mantel.randtest(onion, onion2)$obs
  onion <- dist(potato2_new[, i])
  onion2 <- dist(potato2_new[, i + 1])
  onion4 <- mantel.randtest(onion, onion2)$obs
  mrand_obs <- rbind(mrand_obs, cbind(onion3, onion4))
}

pdf("covid_PCA.pdf")
plot(c(9,10,11,12,13,14,15,16,17,18,19,20,21,22,23,24),
     mrand_obs[, 1], ylim = c(0.8, 1))
lines(spline(x = c(9,10,11,12,13,14,15,16,17,18,19,20,21,22,23,24),
             y = mrand_obs[, 1]))
points(c(9,10,11,12,13,14,15,16,17,18,19,20,21,22,23,24),
       mrand_obs[, 2], ylim = c(0.8, 1), col = "red")
lines(spline(x = c(9,10,11,12,13,14,15,16,17,18,19,20,21,22,23,24),
             y = mrand_obs[, 2]), col = "red")
ElbowPlot(covid)
dev.off()
rm(onion,onion2,onion3,onion4,mrand_obs,potato,potato2, potato_new,potato2_new)

covid <- RunUMAP(covid, dims = 1:20)



saveRDS(refquery, file = "covid_github_replication.RDS")

# ======================================
# Apply metadata
# ======================================
#meta.data <- read.csv("GSE135893_IPF_metadata.csv", header = T, row.names = 1)
#covid@meta.data <- meta.data

#epi_meta <- meta.data[rownames(meta.data) %in% rownames(epi@meta.data), ]
#epi@meta.data <- epi_meta

#endo_meta <- meta.data[rownames(meta.data) %in% rownames(endo@meta.data), ]
#endo@meta.data <- endo_meta


#immune_meta <- meta.data[rownames(meta.data) %in% rownames(immune@meta.data), ]
#immune@meta.data <- immune_meta


#meso_meta <- meta.data[rownames(meta.data) %in% rownames(meso@meta.data), ]
#meso@meta.data <- meso_meta

# ======================================
# Figure 1: B
# ======================================
#DimPlot(covid, group.by = "celltype")

# ======================================
# Figure 1: C
# ======================================
#DimPlot(covid, group.by = "Status")
#DimPlot(covid, group.by = "Diagnosis")
```


```{r}
head(covid)
```
